Created ArticleHeader
This helps with refactoring the layout
