+++ packages/frontend/web/components/ArticleBody.tsx
@@ -2,88 +2,17 @@
 
 import React from 'react';
 import { css, cx } from 'emotion';
+
 import { palette } from '@guardian/pasteup/palette';
-import ClockIcon from '@guardian/pasteup/icons/clock.svg';
-import {
-    from,
-    until,
-    wide,
-    leftCol,
-    desktop,
-    tablet,
-    mobileLandscape,
-} from '@guardian/pasteup/breakpoints';
-import { clearFix, screenReaderOnly } from '@guardian/pasteup/mixins';
-import { headline, textSans, body } from '@guardian/pasteup/typography';
-import { Byline } from '@guardian/guui/components/Byline/Byline';
+import { from, tablet } from '@guardian/pasteup/breakpoints';
+import { headline, textSans } from '@guardian/pasteup/typography';
 import { pillarMap, pillarPalette } from '@frontend/lib/pillars';
 import { ArticleRenderer } from '@frontend/web/components/lib/ArticleRenderer';
-import { ShareCount } from './ShareCount';
+import { getSharingUrls } from '@frontend/model/sharing-urls';
+
 import { SharingIcons } from './ShareIcons';
 import { SubMetaLinksList } from './SubMetaLinksList';
-import { Dateline } from './Dateline';
-
-import { MainMedia } from './MainMedia';
-import { getSharingUrls } from '@frontend/model/sharing-urls';
-import { getAgeWarning } from '@frontend/model/age-warning';
 import { SyndicationButton } from './SyndicationButton';
-import { SeriesSectionLink } from './SeriesSectionLink';
-
-const curly = (x: any) => x;
-
-const wrapper = css`
-    padding-top: 6px;
-    margin-right: 0;
-    margin-left: 0;
-    ${clearFix}
-
-    ${desktop} {
-        max-width: 620px;
-        margin-right: 310px;
-        padding-left: 10px;
-    }
-
-    ${leftCol} {
-        margin-left: 150px;
-        margin-right: 310px;
-        position: relative;
-        :before {
-            content: '';
-            position: absolute;
-            top: 0;
-            bottom: 0;
-            left: 0;
-            height: 100%;
-            width: 1px;
-            background: ${palette.neutral[86]};
-        }
-    }
-
-    ${wide} {
-        margin-left: 230px;
-    }
-
-    header {
-        display: flex;
-        flex-direction: column;
-
-        ${leftCol} {
-            @supports (display: grid) {
-                display: grid;
-                grid-template-areas: 'section headline' 'meta main-media';
-                grid-template-columns: 160px 1fr;
-                margin-left: -160px;
-            }
-        }
-
-        ${wide} {
-            @supports (display: grid) {
-                grid-template-columns: 240px 1fr;
-                margin-left: -240px;
-            }
-        }
-    }
-`;
 
 const pillarColours = pillarMap(
     pillar =>
@@ -128,69 +57,6 @@ const listStyles = css`
     }
 `;
 
-const standfirst = css`
-    ${body(2)};
-    font-weight: 700;
-    color: ${palette.neutral[7]};
-    margin-bottom: 12px;
-
-    ${listStyles};
-
-    p {
-        margin-bottom: 8px;
-    }
-
-    li {
-        ${headline(2)};
-    }
-`;
-
-const standfirstLinks = pillarMap(
-    pillar =>
-        css`
-            a {
-                color: ${pillarPalette[pillar].dark};
-                text-decoration: none;
-                border-bottom: 1px solid ${palette.neutral[86]};
-                transition: border-color 0.15s ease-out;
-            }
-        `,
-);
-
-const leftColWidth = css`
-    ${leftCol} {
-        width: 140px;
-    }
-
-    ${wide} {
-        width: 220px;
-    }
-`;
-
-const headlineCSS = css`
-    @supports (display: grid) {
-        grid-template-areas: 'headline';
-    }
-    ${until.phablet} {
-        padding: 0 10px;
-    }
-`;
-
-const meta = css`
-    ${leftColWidth};
-    @supports (display: grid) {
-        grid-template-areas: 'meta';
-    }
-    ${from.tablet.until.leftCol} {
-        order: 1;
-    }
-
-    ${until.phablet} {
-        padding-left: 10px;
-        padding-right: 10px;
-    }
-`;
-
 const guardianLines = css`
     background-image: repeating-linear-gradient(
         to bottom,
@@ -209,51 +75,6 @@ const guardianLines = css`
 const captionFont = css`
     ${textSans(1)};
     color: ${palette.neutral[46]};
-`;
-
-const mainMedia = css`
-    @supports (display: grid) {
-        grid-template-areas: 'main-media';
-    }
-
-    min-height: 1px;
-    /*
-    Thank you IE11, broken in stasis for all eternity.
-
-    https://github.com/philipwalton/flexbugs/issues/75#issuecomment-161800607
-    */
-
-    margin-bottom: 6px;
-
-    ${until.tablet} {
-        margin: 0;
-        order: -1;
-
-        figcaption {
-            display: none;
-        }
-    }
-
-    img {
-        flex: 0 0 auto; /* IE */
-        width: 100%;
-        height: 100%;
-    }
-
-    figcaption {
-        ${captionFont};
-    }
-`;
-
-const headerStyle = css`
-    ${headline(7)};
-    font-weight: 500;
-    padding-bottom: 24px;
-    padding-top: 3px;
-
-    ${tablet} {
-        padding-bottom: 36px;
-    }
 `;
 
 const bodyStyle = css`
@@ -317,62 +138,6 @@ const linkColour = pillarMap(
     `,
 );
 
-const ageWarningStyle = css`
-    ${textSans(5)};
-    color: ${palette.neutral[7]};
-    background-color: ${palette.highlight.main};
-    display: inline-block;
-    margin-bottom: 6px;
-
-    > strong {
-        font-weight: bold;
-    }
-
-    padding: 6px 10px;
-    margin-top: 6px;
-    margin-left: -10px;
-
-    ${mobileLandscape} {
-        padding-left: 12px;
-    }
-
-    ${tablet} {
-        margin-left: -20px;
-    }
-
-    ${leftCol} {
-        margin-left: -10px;
-        margin-top: -6px;
-        padding-left: 10px;
-    }
-`;
-
-const ageWarningScreenReader = css`
-    ${screenReaderOnly};
-`;
-
-const metaExtras = css`
-    border-top: 1px solid ${palette.neutral[86]};
-    padding-top: 6px;
-    margin-bottom: 6px;
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-
-    ${until.phablet} {
-        margin-left: -10px;
-        margin-right: -10px;
-        padding-left: 10px;
-        padding-right: 10px;
-    }
-`;
-
-const header = css`
-    ${until.phablet} {
-        margin: 0 -10px;
-    }
-`;
-
 const subMeta = css`
     margin-top: 12px;
     padding-top: 18px;
@@ -392,16 +157,6 @@ const subMetaSharingIcons = css`
     }
 `;
 
-const section = css`
-    ${leftColWidth};
-    @supports (display: grid) {
-        grid-template-areas: 'section';
-    }
-    ${until.phablet} {
-        padding: 0 10px;
-    }
-`;
-
 export const ArticleBody: React.FC<{
     CAPI: CAPIType;
     config: ConfigType;
@@ -409,116 +164,59 @@ export const ArticleBody: React.FC<{
     const hasSubMetaSectionLinks = CAPI.subMetaSectionLinks.length > 0;
     const hasSubMetaKeywordLinks = CAPI.subMetaKeywordLinks.length > 0;
     const sharingUrls = getSharingUrls(CAPI.pageId, CAPI.webTitle);
-    const ageWarning = getAgeWarning(CAPI.tags, CAPI.webPublicationDate);
     return (
-        <div className={wrapper}>
-            <header className={header}>
-                <div className={section}>
-                    <SeriesSectionLink CAPI={CAPI} fallbackToSection={true} />
-                </div>
-                <div className={headlineCSS}>
-                    {ageWarning && (
-                        <div className={ageWarningStyle} aria-hidden="true">
-                            <ClockIcon /> This article is more than{' '}
-                            <strong>{ageWarning}</strong>
-                        </div>
-                    )}
-                    <h1 className={headerStyle}>{curly(CAPI.headline)}</h1>
-                    {ageWarning && (
-                        <div className={ageWarningScreenReader}>
-                            This article is more than {` ${ageWarning}`}
-                        </div>
-                    )}
-                    <div
-                        className={cx(standfirst, standfirstLinks[CAPI.pillar])}
-                        dangerouslySetInnerHTML={{
-                            __html: CAPI.standfirst,
-                        }}
-                    />
-                </div>
-                <div className={cx(meta, guardianLines)}>
-                    <Byline
-                        author={CAPI.author}
-                        tags={CAPI.tags}
+        <div>
+            <div
+                className={cx(bodyStyle, linkColour[CAPI.pillar], {
+                    [immersiveBodyStyle]: CAPI.isImmersive,
+                })}
+            >
+                <ArticleRenderer
+                    elements={CAPI.blocks[0] ? CAPI.blocks[0].elements : []}
+                    pillar={CAPI.pillar}
+                    config={CAPI.config}
+                />
+            </div>
+            <div className={cx(subMeta, guardianLines)}>
+                {(hasSubMetaSectionLinks || hasSubMetaKeywordLinks) && (
+                    <span className={subMetaLabel}>Topics</span>
+                )}
+                {hasSubMetaSectionLinks && (
+                    <SubMetaLinksList
+                        links={CAPI.subMetaSectionLinks}
+                        isSectionLinkList={true}
                         pillar={CAPI.pillar}
                     />
-                    <Dateline
-                        dateDisplay={CAPI.webPublicationDateDisplay}
-                        descriptionText={'Published on'}
-                    />
-                    <div className={metaExtras}>
-                        <SharingIcons
-                            sharingUrls={sharingUrls}
-                            pillar={CAPI.pillar}
-                            displayIcons={['facebook', 'twitter', 'email']}
-                        />
-                        <ShareCount config={config} pageId={CAPI.pageId} />
-                    </div>
-                </div>
-                <div className={cx(mainMedia)}>
-                    {CAPI.mainMediaElements.map((element, i) => (
-                        <MainMedia
-                            element={element}
-                            key={i}
-                            pillar={CAPI.pillar}
-                        />
-                    ))}
-                </div>
-            </header>
-
-            <div>
-                <div
-                    className={cx(bodyStyle, linkColour[CAPI.pillar], {
-                        [immersiveBodyStyle]: CAPI.isImmersive,
-                    })}
-                >
-                    <ArticleRenderer
-                        elements={CAPI.blocks[0] ? CAPI.blocks[0].elements : []}
+                )}
+                {hasSubMetaKeywordLinks && (
+                    <SubMetaLinksList
+                        links={CAPI.subMetaKeywordLinks}
+                        isSectionLinkList={false}
                         pillar={CAPI.pillar}
-                        config={CAPI.config}
                     />
-                </div>
-                <div className={cx(subMeta, guardianLines)}>
-                    {(hasSubMetaSectionLinks || hasSubMetaKeywordLinks) && (
-                        <span className={subMetaLabel}>Topics</span>
-                    )}
-                    {hasSubMetaSectionLinks && (
-                        <SubMetaLinksList
-                            links={CAPI.subMetaSectionLinks}
-                            isSectionLinkList={true}
-                            pillar={CAPI.pillar}
-                        />
-                    )}
-                    {hasSubMetaKeywordLinks && (
-                        <SubMetaLinksList
-                            links={CAPI.subMetaKeywordLinks}
-                            isSectionLinkList={false}
-                            pillar={CAPI.pillar}
-                        />
-                    )}
-                    {CAPI.showBottomSocialButtons && (
-                        <SharingIcons
-                            className={subMetaSharingIcons}
-                            sharingUrls={sharingUrls}
-                            pillar={CAPI.pillar}
-                            displayIcons={[
-                                'facebook',
-                                'twitter',
-                                'email',
-                                'linkedIn',
-                                'pinterest',
-                                'whatsApp',
-                                'messenger',
-                            ]}
-                        />
-                    )}
-                    {CAPI.showBottomSocialButtons && (
-                        <SyndicationButton
-                            webUrl={CAPI.webURL}
-                            internalPageCode={CAPI.pageId}
-                        />
-                    )}
-                </div>
+                )}
+                {CAPI.showBottomSocialButtons && (
+                    <SharingIcons
+                        className={subMetaSharingIcons}
+                        sharingUrls={sharingUrls}
+                        pillar={CAPI.pillar}
+                        displayIcons={[
+                            'facebook',
+                            'twitter',
+                            'email',
+                            'linkedIn',
+                            'pinterest',
+                            'whatsApp',
+                            'messenger',
+                        ]}
+                    />
+                )}
+                {CAPI.showBottomSocialButtons && (
+                    <SyndicationButton
+                        webUrl={CAPI.webURL}
+                        internalPageCode={CAPI.pageId}
+                    />
+                )}
             </div>
         </div>
     );
