+++ packages/frontend/web/components/Content.tsx
@@ -3,14 +3,71 @@ import { css } from 'emotion';
 import {
     tablet,
     desktop,
+    leftCol,
     wide,
     palette,
     mobileLandscape,
 } from '@guardian/src-foundations';
+import { clearFix } from '@guardian/pasteup/mixins';
 import { Container } from '@guardian/guui';
 import { ArticleBody } from '@frontend/web/components/ArticleBody';
+import { ArticleHeader } from '@frontend/web/components/ArticleHeader';
 import { AdSlot, labelStyles } from '@frontend/web/components/AdSlot';
 import { namedAdSlotParameters } from '@frontend/model/advertisement';
+
+const wrapper = css`
+    padding-top: 6px;
+    margin-right: 0;
+    margin-left: 0;
+    ${clearFix}
+
+    ${desktop} {
+        max-width: 620px;
+        margin-right: 310px;
+        padding-left: 10px;
+    }
+
+    ${leftCol} {
+        margin-left: 150px;
+        margin-right: 310px;
+        position: relative;
+        :before {
+            content: '';
+            position: absolute;
+            top: 0;
+            bottom: 0;
+            left: 0;
+            height: 100%;
+            width: 1px;
+            background: ${palette.neutral[86]};
+        }
+    }
+
+    ${wide} {
+        margin-left: 230px;
+    }
+
+    header {
+        display: flex;
+        flex-direction: column;
+
+        ${leftCol} {
+            @supports (display: grid) {
+                display: grid;
+                grid-template-areas: 'section headline' 'meta main-media';
+                grid-template-columns: 160px 1fr;
+                margin-left: -160px;
+            }
+        }
+
+        ${wide} {
+            @supports (display: grid) {
+                grid-template-columns: 240px 1fr;
+                margin-left: -240px;
+            }
+        }
+    }
+`;
 
 const secondaryColumn = css`
     position: absolute;
@@ -109,7 +166,10 @@ export const Content = ({ CAPI, config }: Props) => {
                 className={articleContainerStyles}
             >
                 <article className={articleAdStyles}>
-                    <ArticleBody CAPI={CAPI} config={config} />
+                    <div className={wrapper}>
+                        <ArticleHeader CAPI={CAPI} config={config} />
+                        <ArticleBody CAPI={CAPI} config={config} />
+                    </div>
                     <div className={secondaryColumn}>
                         <div className={adSlotWrapper}>
                             <AdSlot
